Move the distribution function interpolation completely to C++
Before, interpolation of values outside the range on momentum grid was handled in Python. This worked OK while Boltzmann integration was also handled there.

But now we do it in C++ and `distribution_interpolation` is called from `integrate` in `integral.cpp`. Hence, in cases when the collision integral requested a value of distribution function at high momenta, C++ did not know how to handle it and freaked out.

That’s why the code worked only when the implementation of binary search was incorrectly handling an argument too large — instead of pointing to the last point on the grid, it was pointing to the second-to-last. Which was slightly incorrect